Changed tracked delivery type to int

+++ Client/Services/SettingsService.cs
@@ -16,7 +16,7 @@ namespace DominosDriverHustleComp.Client.Services
         public float HustleBenchmarkSeconds { get; set; }
         public float OutlierSeconds { get; set; }
         public int MinDels { get; set; }
-        public float MinTrackedPercentage { get; set; }
+        public int MinTrackedPercentage { get; set; }
         public int MaxOverspeeds { get; set; }
 
         public async Task FetchSettings()
