+++ Shared/ViewModels/SettingsVM.cs
@@ -5,7 +5,7 @@
         public float HustleBenchmarkSeconds { get; set; }
         public float OutlierSeconds { get; set; }
         public int MinDels { get; set; }
-        public float MinTrackedPercentage { get; set; }
+        public int MinTrackedPercentage { get; set; }
         public int MaxOverspeeds { get; set; }
     }
 }
